fix: resolve Linux clippy errors (unnecessary_cast, manual_strip)
CI was failing on Linux with 6 clippy errors under -D warnings:
- 4x unnecessary_cast in src/engines/disk/engine.rs (u64 as u64 on
  Linux where statvfs fields are already u64; added #[allow] since
  the cast is needed on macOS where fields are u32)
- 2x manual_strip in src/util/memory.rs (starts_with + manual slice
  replaced with strip_prefix pattern)

+++ src/engines/disk/engine.rs
@@ -96,9 +96,13 @@ fn linux_stats() -> Option<ApfsStats> {
         }
         let buf = buf.assume_init();
 
-        let block_size = buf.f_frsize;
+        #[allow(clippy::unnecessary_cast)]
+        let block_size = buf.f_frsize as u64;
+        #[allow(clippy::unnecessary_cast)]
         let total = buf.f_blocks as u64 * block_size;
+        #[allow(clippy::unnecessary_cast)]
         let free = buf.f_bavail as u64 * block_size;
+        #[allow(clippy::unnecessary_cast)]
         let used = (buf.f_blocks as u64 - buf.f_bfree as u64) * block_size;
 
         // Best-effort: /usr as "system", home as "data"
